Move `TemporaryDirectory` to its own file
Will be reused in another test.
